server/requirements.txt
alembic==1.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
lob==4.0.1
Mako==1.1.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
